+++ Changelog.md
@@ -7,6 +7,8 @@ WhateverGreen Changelog
 - Improved performance with Lilu 1.4.3 APIs
 - Added `-igfxfbdbg` boot argument to debug IGPU framebuffer (debug builds only)
 - Added `igfxagdc=0` boot argument and `disable-agdc` IGPU property to disable AGDC
+- Added `igfxonln=1` boot argument and `force-online` IGPU property force online status for all displays
+- Added `igfxonlnfbs=MASK` boot argument and `force-online-framebuffers` IGPU property to override display status
 
 #### v1.3.7
 - Improved the maximum link rate fix: Now correct the value read from extended DPCD as well. (by @0xFireWolf)